agent_eval/pyproject.toml
[project]
name = "agent-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.36.0",
    "strands-agents-evals>=0.1.15",
    "strands-agents-tools>=0.5.0",
    "openai>=1.0.0",
]
